Make minimum distance surface work in-place to run faster.
Rather than copy the surface of interest, use a pointer to the surface.
Also avoid copies in the piecewise search code where possible.

Avoiding the copy has two effects.  First, it avoids the memory allocation
and free associated with copying significant data.  Second (not obvious
from the local code) is it allows the cached derivative surface to be
calculated once and re-used many times.

This change has significant impact on the remeshing time of CFDMesh.