Cache closed-ness flag for piecewise surfaces
Profiling VSP's CFDMesh showed that the piecewise surface closed tests
were called many times for surfaces that never changed state.

This change introduces a three-state flag.  A surface's u or v closecache
can be UNKNOWN, CLOSED, or OPEN.

This flag is mutable such that changing it 'does not count' in terms of
const.

The flag is reset to UNKNOWN by any method that could change the surface
closed-ness.