Adds get_bounding_box with U/W Limits
This function begins be checking if the input U/W limits are
within the U/W range of the piecewise surface. The surface is then
split at the U/W limits to create patch boundaries at the limits. The
ukey and vkey are then iterated uver to get each patch bounding
box within the input limits.