Add surface Bezier fbatch, f_ubatch, and f_vbatch
These batch calculators process Bezier surface evaluations in batches --
all the evaluations for a given patch in one go.  A surface eval starts
by constructing a curve, then evaluating on that curve.  This change
constructs that curve once and then evaluates it multiple times.

There are two approaches that can be used -- u first or v first.  fbatch
also does a calculation to determine which way is cheaper.  This is based
on an estimate of the cost to build each curve (based on order) and the
number of times a curve will be re-used.