BUGFIX: Bottom Half Super Ellipse Endpoint Slope Index and Inequailty
First, the contorl point index is updated to use the point adjacent
to the endpoint, which controls the slope going into the endpoint.
Next, the inequality for the bottom half is flipped when compared to the
top half, so the y vlalue can't be greater than the max width location.