Add piecewise surface sum.
Most of the code here is in making sure that the parameter maps for the
two surfaces are identical.  This is done by taking the union of the
pmaps and then splitting both surfaces at all those points.

Since these surfaces were generated with a bisection algorithm, the points
should be regularly spaced and we should not run into situations with
arbitrarily spaced pmaps.

This routine should probably be extended to ensure that the minimum and
maximum parameter values for both surfaces are equal.

If other similar operations are extended to piecewise surfaces (from the
base surfaces already covered), then perhaps the pmap compatibility code
should be refactored to be re-used -- thus allowing the sum or product
code to be simplified substantially.

Another option would be a 'simple' version of this routine that assumes
all these checks are satisfied apriori and runs without prepossessing or
diagnostic checks.