Implementation of cylinder projection
Two approaches to projecting a curve onto a cylinder were proposed.
The first simply transforms the start/finish points and derivatives
and constructs cubic segments from that information.  The second scaled
those derivatives such that the curve would lie on the arc of a circle.

In simple tests, it was observed that the first method represented the
projection most accurately, while the second method did a better job of
placing the curve on the cylinder.  A simple average of the two was
proposed.

The more complex approach to projecting a curve onto a cylinder goes
singular when a section of the curve runs parallel to the cylinder axis.

It may be possible to compute a weighted combination where the weight
is derived from a dot product of the segment direction and the cylinder
axis (instaead of the simple average), but we are still playing with a
divide by zero situation.

Better to just use the more simple algorithm and trust the tolerance
to make sure the target accuracy is met.