Eliminate approximate endpoint test in piecewise curves and surfaces
When a curve/surface parameter is close to a segment/patch end, we would
previously force the query point to the endpoint.

While it is important to prevent querying beyond the bounds of a
curve or surface, we should not force points near the endpoints to
be _at_ the endpoints.

This fixes a problem with the cylinder projection unit test and introduces
a problem with the general creator test (which will be fixed in the next
commit).