Fix unit test failure introduced by removing tolerance on curve evaluation.
This test checks whether a generated curve is C2 continuous.  It was only
working before (for float) because the test points were close enough to
the control point that everything was being evaluated at the same place.

When the curves were instead interrogated a finite distance from the
control point, the 2nd derivatives now differed.

This effect is somewhat magnified because small_num is based on the
value of epsion() for data_type.  This means that the test points for
float are further away from the control point.

Consequently, even though C2 is achieved, you get an errant result when
interrogating points 'far' from the control point.

This fix brings the points 100x closer to the control point when data_type
is float.

A better test would be to use a value of small_num that does not vary
with data_type -- thereby allowing true comparisons between different
data_type's.