Rewrite piecewise curve round() to replace more than one segment.
Before, only the segments that use the node being rounded could be
modified.  If a radius was such that the adjacent segment should be dropped
and the curve modification applied to the next curve, the round operation
was simply failed.

This improvement actually simplifies the code quite a bit.  Instead of
pulling out the adjacent segments, the curve is treated as a piecewise
curve the whole time.

The new replace_t method takes care of putting the curve back together.