Make surface-surface intersection shift to point before proceeding.
By shifting both surfaces such that the point is at the origin we're
assured that all coordinates are near zero.  This avoids the potential
floating point problem of subtracting two large nearly identical numbers,
thereby avoiding rounding error.

This does mean that a copy of the surfaces must be made because they are
no longer const.