Modify Newton system solver to improve singular matrix workaround.
The prior code would check for near-zero diagonal entries and use that
condition to turn off certain variables.

This is sensitive to the near-zero tolerance and its interaction with
the tolerance applied by computeInverseWithCheck to determine if a matrix
is invertible.

This change applies the modification to the smallest diagonal entry,
potentially repeating this process until the matrix is invertible.  By
always operating on the smallest (magnitude) diagonal, it eliminates the
role of one of the tolerances.