Fix 'Minimum distance curve g_functor' error message
The finite-difference fallback would sometimes exceed the curve bounds.

Changing the bounds in the if-statement prevents this, using a one-sided
difference if you are near the bounds (not just on or beyond).

Also use a central difference when away from the bounds.

A test was performed revealing that even after this fallback, the
derivative is often zero.  Consequently, the fallback is likely un-helpful
in improving results in practice.