Include some notes (no code changes) about super ellipse clustering.
The referenced paper provides an algorithm for equal arc-length
distribution of points along a super ellipse (with equal exponents).

Their approach provides a local arc length density that can be used to
build points along a line -- but without any information about how
many points there will be in total.  Consequently, an iterative algorithm
would be required to converge on the dS required to yield a pre-specified
N for any given curve.