Eliminate LMN considerations from Code-Eli
Instead of straight-line distance definition of LMN, VSP has moved to
an arc-length definition implemented in VSP (not Code-Eli).

The implementation was moved because computation in Code-Eli requires
the original surface to be kept around - which is not practical given the
surface modifications done by GeomEngine.

Instead, the new implementation extracts an L curve and stores that
so the conversion can be done independent of modifications to the surf.